framework/fit/python/requirements.txt
numpy==1.25.2
PyYAML==6.0.1
requests==2.31.0
tornado==6.3.2